Consolidate log/ files into log/log.go and enhance godocs.
